Extract Gameplay Attributes from HUD Widget (#103)
* New GameplayAttributesWidget

* Remove health/mana code from HUD

Death management is still handled by HUD, therefore HealthC is still needed by it.

* HUD BP update, New widget BP

* Fix: AI moving towards pending kill Actor

* Stop using doubles for mana

There is absolutely no need to use doubles lol, this is due to a dumb mentality I had at the beginning of the project where "the more precision the better".

+++ Source/GameplayUI/Private/AIStatusWidget.cpp
@@ -79,12 +79,12 @@ void UAIStatusWidget::_onDeath(TObjectPtr<AActor> _) {
     _unbindAll();
 }
 
-void UAIStatusWidget::_setCurrentMana(const double newMana) {
+void UAIStatusWidget::_setCurrentMana(const float newMana) {
     _currentMana = newMana;
     _updateMana();
 }
 
-void UAIStatusWidget::_setMaxMana(const double newMaxMana) {
+void UAIStatusWidget::_setMaxMana(const float newMaxMana) {
     _maxMana = newMaxMana;
     _updateMana();
 }
